Refactor ClearlyDefined API integration with dedicated client.
Extracted ClearlyDefined API calls to a dedicated `ClearlyDefinedClient` class, improving modularity and testability. Replaced inlined API logic with a clean interface implementation (`IClearlyDefinedClient`) for better separation of concerns.
